requirements.txt
streamlit>=1.32.0
streamlit-autorefresh>=1.0.1
streamlit-mic-recorder>=0.0.8
openai>=1.30.0
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.2
python-dotenv>=1.0.1
plotly>=5.20.0
scipy>=1.11.0
pytest>=7.4.0
